changing to/from ML Pipline mode in data generator config
