+++ config_data_gen.yml
@@ -1,8 +1,9 @@
 controller: mpc
-PATH_TO_EXPERIMENT_RECORDINGS_DEFAULT: './Experiment_Recordings/'   # Where to save experiment recording per default
+ML_Pipeline_mode: False  # If True a directory structure required by SI_Toolkit will be created, location taken from config_training, not from here.
+split: [0.8, 0.1]  # Train/Val/(Test) split - only matters if you run it in ML Pipeline mode
+PATH_TO_EXPERIMENT_RECORDINGS_DEFAULT: './Experiment_Recordings/'   # Where to save experiment recording per default, can be overwritten in config_training if ML_Pipeline_mode is True
 seed:   # If not assign random seed based on datetime is used
 length_of_experiment: 360  # Length of each experiment in s:
-split: [0.8, 0.1]  # Train/Val/(Test) split - only matters if you run it in ML Pipeline mode
 random_initial_state:
   # Unassigned variables will be randomly initialized (see below)
   position:
